+++ src/lang-xpath/main/uk/co/reecedunn/intellij/plugin/xpath/psi/impl/xpath/XPathAttributeTestPsiImpl.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2016, 2019 Reece H. Dunn
+ * Copyright (C) 2016, 2019-2020 Reece H. Dunn
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,8 +17,14 @@ package uk.co.reecedunn.intellij.plugin.xpath.psi.impl.xpath
 
 import com.intellij.extapi.psi.ASTWrapperPsiElement
 import com.intellij.lang.ASTNode
+import com.intellij.psi.PsiElement
 import uk.co.reecedunn.intellij.plugin.core.data.CacheableProperty
 import uk.co.reecedunn.intellij.plugin.core.sequences.children
+import uk.co.reecedunn.intellij.plugin.intellij.lang.Saxon
+import uk.co.reecedunn.intellij.plugin.intellij.lang.Version
+import uk.co.reecedunn.intellij.plugin.intellij.lang.VersionConformance
+import uk.co.reecedunn.intellij.plugin.intellij.lang.VersionConformanceName
+import uk.co.reecedunn.intellij.plugin.intellij.resources.XPathBundle
 import uk.co.reecedunn.intellij.plugin.xpath.ast.xpath.XPathAttributeTest
 import uk.co.reecedunn.intellij.plugin.xdm.functions.op.op_qname_presentation
 import uk.co.reecedunn.intellij.plugin.xdm.types.XdmAttributeNode
@@ -26,14 +32,28 @@ import uk.co.reecedunn.intellij.plugin.xdm.types.XdmItemType
 import uk.co.reecedunn.intellij.plugin.xdm.types.XdmSequenceType
 import uk.co.reecedunn.intellij.plugin.xdm.types.XsQNameValue
 import uk.co.reecedunn.intellij.plugin.xpath.ast.xpath.XPathTypeName
+import uk.co.reecedunn.intellij.plugin.xpath.ast.xpath.XPathWildcard
 
-class XPathAttributeTestPsiImpl(node: ASTNode) : ASTWrapperPsiElement(node), XPathAttributeTest, XdmItemType {
+class XPathAttributeTestPsiImpl(node: ASTNode) :
+    ASTWrapperPsiElement(node), VersionConformance, VersionConformanceName, XPathAttributeTest, XdmItemType {
     // region ASTDelegatePsiElement
 
     override fun subtreeChanged() {
         super.subtreeChanged()
         cachedTypeName.invalidate()
     }
+
+    // endregion
+    // region VersionConformance
+
+    override val requiresConformance: List<Version>
+        get() = if (conformanceElement is XPathWildcard) listOf(Saxon.VERSION_10_0) else listOf()
+
+    override val conformanceElement: PsiElement
+        get() = children().filterIsInstance<XPathWildcard>().firstOrNull() ?: firstChild
+
+    override val conformanceName: String?
+        get() = XPathBundle.message("construct.wildcard-attribute-test")
 
     // endregion
     // region XPathAttributeTest