+++ src/lang-xpath/test/uk/co/reecedunn/intellij/plugin/xpath/tests/psi/PluginPsiTest.kt
@@ -29,10 +29,7 @@ import uk.co.reecedunn.intellij.plugin.xdm.functions.XdmFunctionReference
 import uk.co.reecedunn.intellij.plugin.xdm.functions.op.op_qname_presentation
 import uk.co.reecedunn.intellij.plugin.xdm.types.*
 import uk.co.reecedunn.intellij.plugin.xpath.ast.plugin.PluginUnionType
-import uk.co.reecedunn.intellij.plugin.xpath.ast.xpath.XPathEQName
-import uk.co.reecedunn.intellij.plugin.xpath.ast.xpath.XPathElementTest
-import uk.co.reecedunn.intellij.plugin.xpath.ast.xpath.XPathFunctionCall
-import uk.co.reecedunn.intellij.plugin.xpath.ast.xpath.XPathTypedMapTest
+import uk.co.reecedunn.intellij.plugin.xpath.ast.xpath.*
 import uk.co.reecedunn.intellij.plugin.xpath.model.getUsageType
 import uk.co.reecedunn.intellij.plugin.xpath.tests.parser.ParserTestCase
 
@@ -240,6 +237,30 @@ private class PluginPsiTest : ParserTestCase() {
                 val type = test as XdmItemType
                 assertThat(type.typeName, `is`("element(*:test)"))
                 assertThat(type.typeClass, `is`(sameInstance(XdmElementNode::class.java)))
+
+                assertThat(type.itemType, `is`(sameInstance(type)))
+                assertThat(type.lowerBound, `is`(1))
+                assertThat(type.upperBound, `is`(1))
+            }
+        }
+    }
+
+    @Nested
+    @DisplayName("XQuery IntelliJ Plugin XPath (2.1.2.8) Attribute Test")
+    internal inner class AttributeTest {
+        @Nested
+        @DisplayName("XQuery IntelliJ Plugin XPath EBNF (30) AttribNameOrWildcard")
+        internal inner class AttribNameOrWildcard {
+            @Test
+            @DisplayName("wildcard")
+            fun wildcard() {
+                val test = parse<XPathAttributeTest>("() instance of attribute ( *:test )")[0]
+                assertThat(op_qname_presentation(test.nodeName!!), `is`("*:test"))
+                assertThat(test.nodeType, `is`(nullValue()))
+
+                val type = test as XdmItemType
+                assertThat(type.typeName, `is`("attribute(*:test)"))
+                assertThat(type.typeClass, `is`(sameInstance(XdmAttributeNode::class.java)))
 
                 assertThat(type.itemType, `is`(sameInstance(type)))
                 assertThat(type.lowerBound, `is`(1))